requirements.txt
torch
transformers
datasets
pyyaml
tqdm
